fix: FSlateFontInfo::Size int32->float compile error on UE 5.7+ (fixes #2)

+++ Source/AutonomixActions/Private/Widget/AutonomixWidgetActions.cpp
@@ -1227,7 +1227,7 @@ FAutonomixActionResult FAutonomixWidgetActions::ExecuteSetWidgetFont(const TShar
 
 	Result.bSuccess = true;
 	Result.ResultMessage = FString::Printf(TEXT("Set font on '%s': size=%d%s%s."),
-		*WidgetName, FontInfo.Size,
+		*WidgetName, static_cast<int32>(FontInfo.Size),
 		!Typeface.IsEmpty() ? *FString::Printf(TEXT(", typeface=%s"), *Typeface) : TEXT(""),
 		!FontFamily.IsEmpty() ? *FString::Printf(TEXT(", family=%s"), *FontFamily) : TEXT(""));
 	Result.ModifiedAssets.Add(AssetPath);
